Generate a compat /1.1 repo with proper redirects
This new scripts duplicates the repo to a /1.1 subdirectory, applies the `to_1_1.ml` opam-admin.top script to it,
and sets redirects to and from /1.1 to handle both <= 1.1 and 1.2 upgrade.

Note that this can be tested before we actually start to make non-1.1 compatible changes to the repo,
it will just set up the redirects and the rewrite will be a no-op.